+++ apps/omsorgsdager-aleneomsorg-dialog/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @navikt/omsorgsdager-aleneomsorg-dialog
 
+## 2.28.45
+
+### Patch Changes
+
+- Updated dependencies
+    - @navikt/sif-common-forms-ds@21.0.41
+    - @navikt/sif-common-soknad-ds@26.1.23
+
 ## 2.28.44
 
 ### Patch Changes