+++ apps/sif-ettersending/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @navikt/sif-ettersending
 
+## 3.49.45
+
+### Patch Changes
+
+- Updated dependencies
+    - @navikt/sif-common-forms-ds@21.0.41
+    - @navikt/sif-common-soknad-ds@26.1.23
+
 ## 3.49.44
 
 ### Patch Changes