+++ apps/ungdomsytelse-deltaker/CHANGELOG.md
@@ -4,10 +4,12 @@
 
 ### Patch Changes
 
-- Oppdatere til gjelderDelerAvMåned flagg på oppgaver for inntekt og kontroll av inntekt
 - Updated dependencies
     - @navikt/ung-deltakelse-opplyser-api-deltaker@0.3.0
     - @navikt/ung-common@0.0.81
+- Updated dependencies
+    - @navikt/sif-common-forms-ds@21.0.41
+    - @navikt/sif-common-soknad-ds@26.1.23
 
 ## 1.5.24